# frameworks/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: To Do list App
  version: 1.0.0
paths:
    /task:
        post:
            description: create a new task
            operationId: createNewTask
            responses:
                '201':
                    description: Task created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaskResponse'
                default:
                    description: Unexpected error
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TaskRequest'
    /task/{id}:
        parameters:
            - in: path
              name: id
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            description: returns a single task from the system
            operationId: getTask
            responses:
                '200':
                    description: Task response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaskResponse'
                '404':
                    description: Task not found
                '400':
                    description: Invalid request
                default:
                    description: Unexpected error
        delete:
            description: returns a single task from the system
            operationId: deleteTask
            responses:
                '204':
                    description: Task deleted
                '404':
                    description: Task not found
                '400':
                    description: Invalid request
                default:
                    description: Unexpected error
        put:
            description: change task status
            operationId: updateTask
            responses:
                '201':
                    description: Task created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaskResponse'
                '204':
                    description: Task updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TaskResponse'
                '400':
                    description: Invalid request
                default:
                    description: Unexpected error
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TaskRequest'
    /tasks:
        get:
            description: returns all tasks from the system
            operationId: getAllTasks
            responses:
                '200':
                    description: A list of tasks
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/TaskResponse'
                '400':
                    description: Invalid request
                default:
                    description: Unexpected error

components:
    schemas:
        TaskResponse:
            type: object
            properties:
                taskResponse:
                    $ref: 'TaskResponse'
        TaskRequest:
            type: object
            properties:
                taskRequest:
                    $ref: 'TaskRequest'